refactor(spp): split fixtures.go monolith by responsibility
Splits the ~683-line build-tagged fixtures.go into three same-package files:
fixtures.go (public types + entry points + proof-system cache),
fixtures_scenario.go (scenarioBuilder construction: scenarios/shapeFlow), and
fixtures_request.go (scenario -> proof request/fixture assembly + field
helpers). The build tag is carried on each file; declarations are byte-identical,
only relocated.
